# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreerSiteRequestDto" : {
        "properties" : {
          "adresseLigne" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "code" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "nom" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "quartier" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "telephone" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "ville" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ModifierSiteRequestDto" : {
        "properties" : {
          "adresseLigne" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "nom" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "quartier" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "telephone" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "ville" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SiteDto" : {
        "properties" : {
          "actif" : {
            "type" : "boolean"
          },
          "adresseLigne" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "etablissementId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "principal" : {
            "type" : "boolean"
          },
          "quartier" : {
            "type" : "string"
          },
          "telephone" : {
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "in" : "header",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API de gestion scolaire Edukeys, éditée par Nova Digital (Togo).",
    "title" : "Edukeys — API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/etablissements/{etablissementId}/sites" : {
      "get" : {
        "operationId" : "listerSites",
        "parameters" : [
          {
            "in" : "path",
            "name" : "etablissementId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SiteDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Sites de l'établissement"
          }
        },
        "summary" : "Liste des sites actifs d'un établissement",
        "tags" : [
          "Sites"
        ]
      },
      "post" : {
        "operationId" : "creerSite",
        "parameters" : [
          {
            "in" : "path",
            "name" : "etablissementId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreerSiteRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiteDto"
                }
              }
            },
            "description" : "Site créé"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiteDto"
                }
              }
            },
            "description" : "Code de site déjà utilisé dans cet établissement"
          }
        },
        "summary" : "Crée un site pour un établissement",
        "tags" : [
          "Sites"
        ]
      }
    },
    "/api/v1/etablissements/{etablissementId}/sites/{siteId}" : {
      "put" : {
        "operationId" : "modifierSite",
        "parameters" : [
          {
            "in" : "path",
            "name" : "etablissementId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          },
          {
            "in" : "path",
            "name" : "siteId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ModifierSiteRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiteDto"
                }
              }
            },
            "description" : "Site modifié"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiteDto"
                }
              }
            },
            "description" : "Site introuvable"
          }
        },
        "summary" : "Modifie un site",
        "tags" : [
          "Sites"
        ]
      }
    },
    "/api/v1/etablissements/{etablissementId}/sites/{siteId}/desactivation" : {
      "post" : {
        "operationId" : "desactiverSite",
        "parameters" : [
          {
            "in" : "path",
            "name" : "etablissementId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          },
          {
            "in" : "path",
            "name" : "siteId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Site désactivé"
          },
          "422" : {
            "description" : "Le site principal ne peut pas être désactivé directement"
          }
        },
        "summary" : "Désactive un site (refusé pour le site principal)",
        "tags" : [
          "Sites"
        ]
      }
    },
    "/api/v1/etablissements/{etablissementId}/sites/{siteId}/principal" : {
      "post" : {
        "operationId" : "designerSitePrincipal",
        "parameters" : [
          {
            "in" : "path",
            "name" : "etablissementId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          },
          {
            "in" : "path",
            "name" : "siteId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Site désigné principal"
          }
        },
        "summary" : "Désigne un site comme principal (bascule transactionnelle)",
        "tags" : [
          "Sites"
        ]
      }
    },
...
  },
  "security" : [
    {
      "bearerAuth" : [ ]
    }
  ],
  "servers" : [
    {
      "description" : "Generated server url",
      "url" : "http://localhost"
    }
  ]
}